src/ashura/requirements.txt
Django==4.1
ipython
requests==2.28.2
psycopg2==2.9.5
confluent-kafka==2.1.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers